refactor: extract ReviewWorkspaceTimelineEntryKind.swift from ReviewWorkspaceTypes.swift
Splits the `ReviewWorkspaceTimelineEntryKind` enum (transcript /
marker / screenshot cases used by the timeline entry) out of
ReviewWorkspaceTypes.swift.

Byte-identical move of lines 40-56. Both files import Foundation.

ReviewWorkspaceTypes.swift: 73 → 56 lines. Tests: 667.
