+++ Sources/BugNarrator/Utilities/ReviewWorkspaceTypes.swift
@@ -37,23 +37,6 @@ struct ReviewWorkspaceTimelineEntry: Identifiable, Equatable {
     }
 }
 
-enum ReviewWorkspaceTimelineEntryKind: Equatable {
-    case transcript
-    case marker
-    case screenshot
-
-    var sortPriority: Int {
-        switch self {
-        case .transcript:
-            return 2
-        case .marker:
-            return 0
-        case .screenshot:
-            return 1
-        }
-    }
-}
-
 struct ReviewSummaryIssueGroup: Equatable {
     let category: ExtractedIssueCategory
     let issues: [ExtractedIssue]
